Breakout HTTP helper functions
