+++ esp32/src/code.py
@@ -2,9 +2,8 @@ import board
 import wifi
 import neopixel
 import socketpool
-import io
-import re
 from digitalio import DigitalInOut, Direction
+import ampule
 
 try:
     from secrets import secrets
@@ -31,91 +30,19 @@ pool = socketpool.SocketPool(wifi.radio)
 socket = pool.socket()
 socket.bind(['0.0.0.0', 80])
 socket.listen(1)
-
 print('Listening on 0.0.0.0:80')
 
-routes = []
-variable_re = re.compile("^<([a-zA-Z]+)>$")
-
-def get_request(client):
-    try:
-        client.setblocking(False)
-        buffer = bytearray(1024)
-        client.recv_into(buffer)
-        reader = io.BytesIO(buffer)
-    except OSError:
-        return None
-
-    line = str(reader.readline(), "utf-8")
-    (method, full_path, version) = line.rstrip("\r\n").split(None, 2)
-    path = full_path.split("?")[0]
-    query_string = full_path.split("?")[1]
-
-    param_list = query_string.split("&")
-    params = {}
-    for param in param_list:
-        key_val = param.split("=")
-        if len(key_val) == 2:
-            params[key_val[0]] = key_val[1]
-
-    headers = {}
-    for line in reader:
-        if line == b'\r\n': break
-        header = str(line, "utf-8")
-        title, content = header.split(":", 1)
-        headers[title.strip().lower()] = content.strip()
-
-    data = ""
-    for line in reader:
-        if line == b'\r\n': break
-        data += str(line, "utf-8")
-
-    return (method, path, params, headers, data)
-
-def send_response(client, code, headers, data):
-    response = "HTTP/1.1 %i\r\n" % code
-
-    headers["server"] = "esp32server"
-    headers["connection"] = "close"
-    for k, v in headers.items():
-        response += "%s: %s\r\n" % (k, v)
-
-    response += data
-    response += "\r\n"
-
-    client.send(response.encode("utf-8"))
-    client.close()
-
-def on_request(rule, request_handler):
-    regex = "^"
-    rule_parts = rule.split("/")
-    for part in rule_parts:
-        var = variable_re.match(part)
-        if var:
-            regex += r"([a-zA-Z0-9_-]+)\/"
-        else:
-            regex += part + r"\/"
-    regex += "?$"
-    routes.append(
-        (re.compile(regex), {"func": request_handler})
-    )
-
-def route(rule):
-    return lambda func: on_request(rule, func)
-
-def match_route(path):
-    for matcher, route in routes:
-        match = matcher.match(path)
-        if match:
-            return (match.groups(), route)
-    return None
-
-@route("/led_on")
+@ampule.route("/led_on")
 def led_on(request):
-    print("led on!")
     pixels.fill((128, 64, 128))
     pixels.show()
-    return ("200 OK", [], "led on!")
+    return (200, {"content-type": "application/json"}, '{success: true}')
+
+@ampule.route("/led_off")
+def led_on(request):
+    pixels.fill(0)
+    pixels.show()
+    return (200, {"content-type": "application/json"}, '{success: true}')
 
 pixels.fill((16, 64, 128))
 pixels.show()
@@ -124,15 +51,13 @@ print("Pixels enabled.")
 while True:
     client, remote_address = socket.accept()
     print("Received request from:", remote_address)
-    request = get_request(client)
+    request = ampule.get_request(client)
     if request:
         (method, path, params, headers, data) = request
-        print(path)
-        match = match_route(path)
+        print("Matching path: %s" % path)
+        match = ampule.match_route(path)
         if match:
             args, route = match
-            status, headers, resp_data = route["func"](request, *args)
+            status, headers, body = route["func"](request, *args)
 
-    headers = {"content-type": "application/json"}
-    body = '{success: true}'
-    send_response(client, 200, headers, body)
+    ampule.send_response(client, status, headers, body)
